Use a byte buffer for audio to minimize buffer copy's

+++ moonlight-common/src/com/limelight/nvstream/av/audio/AudioRenderer.java
@@ -3,7 +3,7 @@ package com.limelight.nvstream.av.audio;
 public interface AudioRenderer {
 	public void streamInitialized(int channelCount, int sampleRate);
 	
-	public void playDecodedAudio(short[] audioData, int offset, int length);
+	public void playDecodedAudio(byte[] audioData, int offset, int length);
 	
 	public void streamClosing();
 }
